About to fix the communication between the out_packet in STA.h and in_packet in Channel.h

+++ Aux.h
@@ -6,7 +6,7 @@ struct Packet
 	int source;
 	int destination;
 	int L;
-	int seq;
+	long int seq;
 	double tx_time; //time at which a packet is transmitted
 	double queuing_time; //time at which the packet is deposited at end of the queue
 	double contention_time;
